+++ data-browser/src/views/FolderPage/GridItem/FileGridItem.tsx
@@ -1,6 +1,9 @@
-import { properties, useString } from '@tomic/react';
 import React from 'react';
 import styled from 'styled-components';
+import { useFileInfo } from '../../../hooks/useFile';
+import { useFilePreviewSizeLimit } from '../../../hooks/useFilePreviewSizeLimit';
+import { isTextFile } from '../../File/isTextFile';
+import { TextPreview } from '../../File/TextPreview';
 import { InnerWrapper } from './components';
 import { GridItemViewProps } from './GridItemViewProps';
 
@@ -14,15 +17,23 @@ const imageMimeTypes = new Set([
 ]);
 
 export function FileGridItem({ resource }: GridItemViewProps): JSX.Element {
-  const [fileUrl] = useString(resource, properties.file.downloadUrl);
-  const [mimetype] = useString(resource, properties.file.mimetype);
+  const { downloadUrl, mimeType, bytes } = useFileInfo(resource);
+  const previewSizeLimit = useFilePreviewSizeLimit();
 
-  if (imageMimeTypes.has(mimetype!)) {
+  if (bytes >= previewSizeLimit) {
+    return <TextWrapper>To large for preview</TextWrapper>;
+  }
+
+  if (imageMimeTypes.has(mimeType)) {
     return (
       <InnerWrapper>
-        <Image src={fileUrl} alt='' />
+        <Image src={downloadUrl} alt='' loading='lazy' />
       </InnerWrapper>
     );
+  }
+
+  if (isTextFile(mimeType)) {
+    return <StyledTextPreview downloadUrl={downloadUrl} mimeType={mimeType} />;
   }
 
   return <TextWrapper>No preview available</TextWrapper>;
@@ -39,4 +50,14 @@ const TextWrapper = styled(InnerWrapper)`
   display: grid;
   place-items: center;
   color: ${p => p.theme.colors.textLight};
+`;
+
+const StyledTextPreview = styled(TextPreview)`
+  padding: ${p => p.theme.margin}rem;
+  color: ${p => p.theme.colors.textLight};
+
+  &:is(pre) {
+    padding: 0;
+    padding-inline: ${p => p.theme.margin}rem;
+  }
 `;